Add schema module and model
